Implement products business logics

+++ src/main/java/org/example/lookinsure/service/dto/ProductReviewDTO.java
@@ -12,6 +12,6 @@ public class ProductReviewDTO {
     private Long productId;
     private List<CommentDTO> comments;
     private Double averageRate;
-    private Long totalReviews;
+    private Integer totalReviews;
 
 }
